requirements.txt
anndata~=0.7.8
h5py~=3.6.0
numpy~=1.21.4
scipy~=1.7.2
matplotlib~=3.5.0
seaborn~=0.11.2
ConfigArgParse~=1.5.3
tqdm~=4.62.3
scikit-learn~=1.0.1
bio~=1.3.2
pysam~=0.18.0
tensorboard~=2.7.0
scanpy==1.8.2
jupyter~=1.0.0
icecream~=2.1.1
shap~=0.40.0
setuptools==59.5.0
torchinfo~=1.5.4
leidenalg